# Optimize merkledb hashing