## Why this should be merged

Before:
```
Benchmark_CalculateID/empty_node-12         	 5420020	       219.5 ns/op	     144 B/op	       1 allocs/op
Benchmark_CalculateID/has_value-12          	 5191927	       233.1 ns/op	     144 B/op	       1 allocs/op
Benchmark_CalculateID/1_child-12            	 3997491	       297.8 ns/op	     177 B/op	       2 allocs/op
Benchmark_CalculateID/2_children-12         	 3156843	       379.3 ns/op	     210 B/op	       2 allocs/op
Benchmark_CalculateID/16_children-12        	  815900	      1515   ns/op	     720 B/op	       2 allocs/op
```

After:
```
Benchmark_CalculateID/empty_node-12         	11578461	       101.8 ns/op	       0 B/op	       0 allocs/op
Benchmark_CalculateID/has_value-12          	10477825	       113.2 ns/op	       0 B/op	       0 allocs/op
Benchmark_CalculateID/1_child-12            	 7845493	       155.4 ns/op	       0 B/op	       0 allocs/op
Benchmark_CalculateID/2_children-12         	 5576605	       215.0 ns/op	       0 B/op	       0 allocs/op
Benchmark_CalculateID/16_children-12        	 1000000	      1009   ns/op	       0 B/op	       0 allocs/op
```

## How this works

Allocs currently happen in 2 places.
1. For the `bytes.Buffer`
2. For the `maps.Keys`

- By writing directly into the `sha256` writer, we can avoid allocating the `bytes.Buffer`.
- By allocating a slice large enough for the biggest supported branch factor, we can avoid allocating the key slice on the heap (it is placed on the stack).

## How this was tested

- [X] Added benchmark
- [X] CI