@patrick-ogrady 2024-03-29
On my benchmarks, I found this improved performance by ~10-20%.